Migrate editor_ui to editor
