+++ CMakeLists.txt
@@ -63,7 +63,6 @@ set(CORE_SRC
 
 set(EDITOR_SRC
     src/editor/editor.cpp
-    src/editor/ui/editor_ui.cpp
     src/editor/ui/log_window.cpp
     src/editor/ui/main_menu_bar.cpp
     src/editor/ui/scene_graph_window.cpp
